requirements.txt
absl-py==2.3.1
addict==2.4.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
basicsr @ git+https://github.com/XPixelGroup/BasicSR.git@8d56e3a045f9fb3e1d8872f92ee4a4f07f886b0a
beautifulsoup4==4.14.2
boto3==1.40.64
botocore==1.40.64
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
facexlib==0.3.0
ffmpeg-python==0.2.0
filelock==3.20.0
filetype==1.2.0
filterpy==1.4.5
fire==0.7.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
future==1.0.0
gfpgan==1.3.8
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
huggingface-hub==0.36.0
idna==3.7
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.3.0
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.9
lazy_loader==0.4
llvmlite==0.45.1
lmdb==1.7.5
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.10.7
moviepy==2.2.1
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
numba==0.62.1
numpy==2.2.6
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pi_heif==1.1.1
pillow==11.3.0
pillow-avif-plugin==1.5.2
platformdirs==4.5.0
playwright==1.55.0
pluggy==1.6.0
polars==1.35.1
polars-runtime-32==1.35.1
proglog==0.1.12
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.2
pybboxes==0.1.6
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pyee==13.0.0
Pygments==2.19.2
pyparsing==3.2.5
PySocks==1.7.1
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
realesrgan @ git+https://github.com/xinntao/Real-ESRGAN@a4abfb2979a7bbff3f69f58f58ae324608821e27
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
roboflow==1.2.11
s3transfer==0.14.0
safetensors==0.6.2
sahi==0.11.36
scikit-image==0.25.2
scipy==1.16.3
seaborn==0.13.2
selenium==4.38.0
shapely==2.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SpeechRecognition==3.14.3
sympy==1.14.0
tb-nightly==2.21.0a20251023
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.2.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
tifffile==2025.10.16
tokenizers==0.22.1
torch==2.9.0+cu128
torchaudio==2.9.0+cu128
torchvision==0.24.0+cu128
tqdm==4.67.1
transformers==4.57.1
trio==0.31.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.223
ultralytics-thop==2.0.18
urllib3==2.5.0
uv==0.9.5
wcwidth==0.2.14
websocket-client==1.9.0
Werkzeug==3.1.3
wsproto==1.2.0
yapf==0.43.0
yarl==1.22.0
yolov5==7.0.14
yolov8==0.0.2
